Move spherical harmonics to hoomd_order::math
Refactor the API for consistency with the rest of the code base.
